fix: add the check for remove file and add the duration message to console

+++ cmd/uuid/main.go
@@ -5,6 +5,7 @@ import (
 	"github.com/AndreiZernov/learn_go_with_saltpay_exercise_one/adapters/files"
 	"os"
 	"strconv"
+	"time"
 )
 
 const envAuthKeysPathname = "AUTH_KEYS_PATHNAME"
@@ -13,13 +14,17 @@ const keysNotGeneratedErrorMessage = "Keys was not generated, please specify the
 func main() {
 	var (
 		err              error
+		startTime        = time.Now()
 		toGetAllArgs     = os.Args[1:]
 		authKeysPathname = os.Getenv(envAuthKeysPathname)
 	)
 
 	if len(toGetAllArgs) > 0 {
 		number, _ := strconv.Atoi(toGetAllArgs[0])
-		files.RemoveFile(authKeysPathname)
+		_, err = os.Stat(authKeysPathname)
+		if err == nil {
+			files.RemoveFile(authKeysPathname)
+		}
 		for i := 0; i < number; i++ {
 			err = files.UUIDGenerator()
 		}
@@ -27,8 +32,11 @@ func main() {
 			fmt.Println(err.Error())
 		} else {
 			fmt.Printf("Successfully generated %d uuid keys in %s \n", number, authKeysPathname)
+			seconds := int(time.Since(startTime) / time.Second)
+			fmt.Printf("To generate %s keys it took %d Seconds \n", toGetAllArgs[0], seconds)
 		}
 	} else {
 		fmt.Printf(keysNotGeneratedErrorMessage)
 	}
+
 }
